requirements.txt
Django==4.2
django-allauth==65.0.2
qrcode==8.0
djangorestframework==3.15.2
Pillow==11.0.0